Move combine cubes operator into misc, and simplify
