+++ src/CSET/operators/_utils.py
@@ -117,54 +117,3 @@ def is_transect(cube: iris.cube.Cube) -> bool:
 
     # Passed criteria so return True
     return True
-
-
-def combine_cubes_into_cubelist(
-    first: iris.cube.Cube | iris.cube.CubeList, **kwargs
-) -> iris.cube.CubeList:
-    """Operator that combines multiple cubes or cubelists into one.
-
-    Arguments
-    ---------
-    first: iris.cube | iris.cube.CubeList
-        First cube or cubelist to merge into cubelist.
-    second: iris.cube | iris.cube.CubeList
-        Second cube or cubelist to merge into cubelist. This must be a named argument.
-    third: iris.cube | iris.cube.CubeList
-        There can be any number of additional constraint, they just need unique
-        names.
-    ...
-
-    Returns
-    -------
-    combined_cubelist: iris.cube.CubeList
-        Combined cubelist containing all cubes/cubelists.
-
-    Raises
-    ------
-    TypeError
-    If the provided arguments are not either an iris.cube or iris.cube.CubeList.
-    """
-    # Create empty cubelist to store cubes/cubelist
-    all_cubes = iris.cube.CubeList()
-
-    # For first argument...
-    if type(first) == iris.cube.Cube:
-        all_cubes.append(first)
-    elif type(first) == iris.cube.CubeList:
-        for cube in first:
-            all_cubes.append(cube)
-    else:
-        raise TypeError("Not a cube or cubelist!")
-
-    # For all subsequent arguments
-    for item in kwargs.values():
-        if type(item) == iris.cube.Cube:
-            all_cubes.append(item)
-        elif type(item) == iris.cube.CubeList:
-            for cube in item:
-                all_cubes.append(cube)
-        else:
-            raise TypeError("Not a cube or cubelist!")
-
-    return all_cubes
