+++ tests/operators/test_utils.py
@@ -76,60 +76,3 @@ def test_is_transect_correctcoord(transect_source_cube):
     # Retain only time and latitude coordinate, so it passes the first spatial coord test.
     transect_source_cube_slice = transect_source_cube[:, :, :, 0]
     assert common_operators.is_transect(transect_source_cube_slice)
-
-
-def test_combine_singlecube_into_cubelist(source_cube):
-    """Test case of single cube into cubelist."""
-    cubelist = common_operators.combine_cubes_into_cubelist(source_cube)
-
-    expected_cubelist = (
-        "[<iris 'Cube' of surface_altitude / (m) (latitude: 200; longitude: 200)>]"
-    )
-    assert repr(cubelist) in expected_cubelist
-
-
-def test_combine_singlecubelist_into_cubelist(source_cube, transect_source_cube):
-    """Test case of single cubelist into cubelist."""
-    cubelist = common_operators.combine_cubes_into_cubelist(
-        iris.cube.CubeList([source_cube, transect_source_cube])
-    )
-
-    expected_cubelist = "[<iris 'Cube' of surface_altitude / (m) (latitude: 200; longitude: 200)>,\n<iris 'Cube' of air_temperature / (K) (time: 2; pressure: 16; latitude: 6; longitude: 6)>]"
-    assert repr(cubelist) in expected_cubelist
-
-
-def test_combine_single_noncompliant_into_cubelist():
-    """Test case of single object which isn't cube or cubelist."""
-    with pytest.raises(TypeError):
-        common_operators.combine_cubes_into_cubelist("hello")
-
-
-def test_combine_multiplecube_into_cubelist(source_cube, transect_source_cube):
-    """Test case of multiple cube into cubelist."""
-    cubelist = common_operators.combine_cubes_into_cubelist(
-        source_cube, a=transect_source_cube
-    )
-
-    expected_cubelist = "[<iris 'Cube' of surface_altitude / (m) (latitude: 200; longitude: 200)>,\n<iris 'Cube' of air_temperature / (K) (time: 2; pressure: 16; latitude: 6; longitude: 6)>]"
-
-    assert repr(cubelist) in expected_cubelist
-
-
-def test_combine_multiple_cube_and_noncompliant_into_cubelist(source_cube):
-    """Test case of a valid cube with some non compliant data which isn't cube or cubelist."""
-    with pytest.raises(TypeError):
-        common_operators.combine_cubes_into_cubelist(source_cube, a="hello")
-
-
-def test_combine_multiplecube_mixed_into_cubelist(source_cube, transect_source_cube):
-    """Test case of multiple cubes and cubelist into cubelist."""
-    cubelist = common_operators.combine_cubes_into_cubelist(
-        source_cube, a=transect_source_cube
-    )
-    out_cubelist = common_operators.combine_cubes_into_cubelist(
-        source_cube, a=cubelist, b=transect_source_cube
-    )
-
-    expected_cubelist = "[<iris 'Cube' of surface_altitude / (m) (latitude: 200; longitude: 200)>,\n<iris 'Cube' of surface_altitude / (m) (latitude: 200; longitude: 200)>,\n<iris 'Cube' of air_temperature / (K) (time: 2; pressure: 16; latitude: 6; longitude: 6)>,\n<iris 'Cube' of air_temperature / (K) (time: 2; pressure: 16; latitude: 6; longitude: 6)>]"
-
-    assert repr(out_cubelist) in expected_cubelist
